WIP: Rework and modernize builds (#49)
## Which problem is this PR solving?

We need to be able to upload builds to S3 as well as Github, because we have links to the builds that people build into build scripts. This restores that capability and modernizes it. 

## Short description of the changes

- Clean up the CircleCI config and the shell script that builds packages. 

This PR will be followed up with further doc work to explain how to access the images.